requirements.txt
fastapi~=0.86.0
fastapi-sqlalchemy
psycopg2-binary
pydantic[dotenv]~=1.10.2
uvicorn~=0.19.0
alembic~=1.8.1
SQLAlchemy~=1.4.43
gunicorn

python-dotenv~=0.21.0
Faker~=15.3.1
starlette~=0.20.4